requirements.txt
requests==2.32.5
python-dotenv
pymysql
